backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
bcrypt==4.0.1
python-jose[cryptography]
python-multipart